extract some code into utils
